replaced the double data type with int, for the test cases proper swap look

+++ B180093CS_1/Q2.java
@@ -2,12 +2,12 @@ import java.util.*;
 
 class Q2 {
     public static void main(String args[]) {
-        double a, b;
+        int a, b;
         Scanner scan = new Scanner(System.in);
         System.out.println("a: ");
-        a = scan.nextDouble();
+        a = scan.nextInt();
         System.out.println("b: ");
-        b = scan.nextDouble();
+        b = scan.nextInt();
         a = a + b;
         b = a - b;
         a = a - b;
